+++ src/sound.cpp
@@ -214,7 +214,7 @@ static void sound_callback (void *userdata, Uint8 *stream, int len)
     in_callback = 1;
     if (! closing_sound) {
 #ifdef USE_SOUND_SEMS
-         while ((!closing_sound) && (((unsigned)render_sndbuff)==((unsigned)callback_sndbuff)))
+         while ((!closing_sound) && (((unsigned long)render_sndbuff)==((unsigned long)callback_sndbuff)))
 	 		uae_sem_wait (&data_available_sem);
 #else
 	n_callback_sndbuff=(n_callback_sndbuff+1)&7;
@@ -231,7 +231,7 @@ static void sound_callback (void *userdata, Uint8 *stream, int len)
 	}
 #endif
 #ifdef USE_SOUND_SEMS
-	 if (((unsigned)render_sndbuff)!=((unsigned)callback_sndbuff)) 
+	 if (((unsigned long)render_sndbuff)!=((unsigned long)callback_sndbuff))
 #endif
 	 {
 #ifndef DREAMCAST
